Intento Facturacion/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pyodbc
python-multipart
jinja2
python-dotenv
pydantic
alembic
fastapi-mail
Pillow
aiofiles
reportlab
